squashing requests package
