--- internal/requests/aad_instance_discovery_test.go
@@ -1,35 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-package requests
-
-import (
-	"context"
-	"reflect"
-	"testing"
-
-	"github.com/AzureAD/microsoft-authentication-library-for-go/internal/msalbase"
-)
-
-func TestGetMetadataEntry(t *testing.T) {
-	authInfo := msalbase.AuthorityInfo{
-		Host: "login.microsoft.com",
-	}
-	mockWRM := new(MockWebRequestManager)
-	metEntry := InstanceDiscoveryMetadata{
-		Aliases: []string{"login.microsoft.com"},
-	}
-	instanceDisc := CreateAadInstanceDiscovery(mockWRM)
-	instanceResp := InstanceDiscoveryResponse{
-		TenantDiscoveryEndpoint: "",
-		Metadata:                []InstanceDiscoveryMetadata{metEntry},
-	}
-	mockWRM.On("GetAadinstanceDiscoveryResponse", authInfo).Return(instanceResp, nil)
-	actualMet, err := instanceDisc.GetMetadataEntry(context.Background(), authInfo)
-	if err != nil {
-		t.Errorf("Error should be nil, but it is %v", err)
-	}
-	if !reflect.DeepEqual(actualMet, metEntry) {
-		t.Errorf("Actual metadata entry %+v differs from expected metadata entry %+v", actualMet, metEntry)
-	}
-}
